refactor: extract SearchPage cards + formatters (F5)
Extract the four presentational result cards (StockOverviewCard,
KeyMetricsCard, AnalystRatingsCard, StockNewsCard) into
components/search/StockCards.js and the pure display formatters into
components/search/format.js. SearchPage keeps its fetch/state logic and
search-specific helpers and drops to      913 lines. Pure code movement,
guarded by SearchPage's existing 6-test suite (all green); lint + build
clean.
